LayoutTests/imported/w3c/web-platform-tests/html/browsers/browsing-the-web/back-forward-cache/events-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: Invalid URL

Harness Error (TIMEOUT), message = null

NOTRUN SameOrigin Should be BFCached but actually wasn't
NOTRUN SameSite Should be BFCached but actually wasn't
TIMEOUT CrossSite Test timed out
NOTRUN beforeunload
NOTRUN unload
